packages/ingestion/requirements.txt
# Core dependencies
fastapi>=0.104.0
pydantic>=2.4.0
asyncio>=3.4.3
aiohttp>=3.8.5

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0
google-cloud-pubsub>=2.18.0
google-cloud-aiplatform>=1.31.0

# Database dependencies
asyncpg>=0.28.0
pgvector>=0.2.0

# Text extraction dependencies
pypdf>=3.16.0
pymupdf>=1.22.5  # Optional, better PDF extraction
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.1.0
extract-msg>=0.41.1  # For Outlook MSG files
python-magic>=0.4.27  # For file type detection

# File processing dependencies
zipfile36>=0.1.3  # For ZIP file handling
tenacity>=8.2.3  # For retry logic

# Optional dependencies (uncomment if needed)
# google-cloud-vision>=3.4.3  # For OCR on images
# camelot-py>=0.11.0  # For table extraction from PDFs
# html2text>=2020.1.16  # For HTML to text conversion
# beautifulsoup4>=4.12.2  # For HTML parsing
